refactor(scan): extract CLI rendering into src/cli/render-*.ts files
Splits the 899-line src/scan.ts god-orchestrator: keeps the actual scan
flow in scan.ts (313 lines) and lifts the CLI rendering helpers into
6 focused files under src/cli/.

scan.ts before: project detection prints, diagnostic group formatters,
score bar + face renderer, summary + share URL builder, tmpdir
diagnostics dump, Node-for-oxlint resolution prompt -- all interleaved
with the scan orchestrator.

scan.ts after: just the orchestrator -- option merge, runScan flow,
lint/dead-code Promise.all, score calculation, calls to the renderers.

Files extracted:
  src/cli/render-diagnostics.ts        317 LOC  printDiagnostics +
                                                category/rule group
                                                formatters + verbose
                                                map + helpers
  src/cli/render-score-header.ts        74 LOC  printScoreHeader /
                                                printBrandingOnlyHeader /
                                                printNoScoreHeader (+
                                                doctor face + score bar)
  src/cli/render-summary.ts             82 LOC  printSummary +
                                                printCountsSummaryLine +
                                                buildShareUrl
  src/cli/render-project-detection.ts   50 LOC  printProjectDetection
                                                (spinner-driven "Found
                                                React 19" lines)
  src/cli/write-diagnostics-directory.ts 27 LOC writeDiagnosticsDirectory
                                                (the /tmp dump)
  src/cli/resolve-oxlint-node.ts        66 LOC  resolveOxlintNode (Node
                                                version detection +
                                                "install via nvm?"
                                                prompt)

No behavior change. Same orchestration, same output, same scan() signature.
Each extracted file only imports what it actually uses (no barrel re-exports).

Tests: 734 passing, typecheck/lint/format clean, build produces identical
dist/cli.js.
