+++ packages/react-doctor/src/cli/write-diagnostics-directory.ts
@@ -1,0 +1,27 @@
+import { randomUUID } from "node:crypto";
+import { mkdirSync, writeFileSync } from "node:fs";
+import { tmpdir } from "node:os";
+import { join } from "node:path";
+import type { Diagnostic } from "../types.js";
+import { groupBy } from "../utils/group-by.js";
+import { formatRuleSummary, sortRuleGroupsByImportance } from "./render-diagnostics.js";
+
+export const writeDiagnosticsDirectory = (diagnostics: Diagnostic[]): string => {
+  const outputDirectory = join(tmpdir(), `react-doctor-${randomUUID()}`);
+  mkdirSync(outputDirectory, { recursive: true });
+
+  const ruleGroups = groupBy(
+    diagnostics,
+    (diagnostic) => `${diagnostic.plugin}/${diagnostic.rule}`,
+  );
+  const sortedRuleGroups = sortRuleGroupsByImportance([...ruleGroups.entries()]);
+
+  for (const [ruleKey, ruleDiagnostics] of sortedRuleGroups) {
+    const fileName = ruleKey.replace(/\//g, "--") + ".txt";
+    writeFileSync(join(outputDirectory, fileName), formatRuleSummary(ruleKey, ruleDiagnostics));
+  }
+
+  writeFileSync(join(outputDirectory, "diagnostics.json"), JSON.stringify(diagnostics));
+
+  return outputDirectory;
+};
