+++ packages/react-doctor/src/scan.ts
@@ -1,580 +1,39 @@
-import { randomUUID } from "node:crypto";
-import { mkdirSync, writeFileSync } from "node:fs";
-import { tmpdir } from "node:os";
-import { join } from "node:path";
 import { performance } from "node:perf_hooks";
 import {
-  MAX_CATEGORY_GROUPS_SHOWN_NON_VERBOSE,
-  MAX_RULE_GROUPS_PER_CATEGORY_NON_VERBOSE,
-  MILLISECONDS_PER_SECOND,
+  ERROR_RULE_PENALTY,
   OFFLINE_MESSAGE,
   OXLINT_NODE_REQUIREMENT,
-  ERROR_RULE_PENALTY,
-  OXLINT_RECOMMENDED_NODE_MAJOR,
-  OUTPUT_DETAIL_WRAP_WIDTH_CHARS,
   PERFECT_SCORE,
-  RULE_NAME_COLUMN_WIDTH_CHARS,
-  SCORE_BAR_WIDTH_CHARS,
-  SCORE_GOOD_THRESHOLD,
-  SCORE_OK_THRESHOLD,
-  SHARE_BASE_URL,
   WARNING_RULE_PENALTY,
 } from "./constants.js";
+import {
+  printBrandingOnlyHeader,
+  printScoreHeader,
+  printNoScoreHeader,
+} from "./cli/render-score-header.js";
+import { printDiagnostics } from "./cli/render-diagnostics.js";
+import { printProjectDetection } from "./cli/render-project-detection.js";
+import { printSummary } from "./cli/render-summary.js";
+import { resolveOxlintNode } from "./cli/resolve-oxlint-node.js";
 import { NoReactDependencyError } from "./errors.js";
-import { resolveConfigRootDir } from "./utils/resolve-config-root-dir.js";
-import type {
-  Diagnostic,
-  ProjectInfo,
-  ReactDoctorConfig,
-  ScanOptions,
-  ScanResult,
-  ScoreResult,
-} from "./types.js";
-import { buildHiddenDiagnosticsSummary } from "./utils/build-hidden-diagnostics-summary.js";
+import type { Diagnostic, ReactDoctorConfig, ScanOptions, ScanResult } from "./types.js";
 import {
   calculateScore,
   calculateScoreBreakdown,
   calculateScoreLocally,
 } from "./utils/calculate-score.js";
-import { colorizeByScore } from "./utils/colorize-by-score.js";
 import { combineDiagnostics } from "./utils/combine-diagnostics.js";
-import { computeJsxIncludePaths } from "./utils/jsx-include-paths.js";
-import { discoverProject, formatFrameworkName } from "./utils/discover-project.js";
+import { discoverProject } from "./utils/discover-project.js";
 import { formatErrorChain } from "./utils/format-error-chain.js";
-import { groupBy } from "./utils/group-by.js";
 import { highlighter } from "./utils/highlighter.js";
-import { indentMultilineText } from "./utils/indent-multiline-text.js";
-import { toRelativePath } from "./utils/to-relative-path.js";
+import { computeJsxIncludePaths } from "./utils/jsx-include-paths.js";
 import { loadConfigWithSource } from "./utils/load-config.js";
 import { isLoggerSilent, logger, setLoggerSilent } from "./utils/logger.js";
-import { prompts } from "./utils/prompts.js";
-import { wrapIndentedText } from "./utils/wrap-indented-text.js";
-import {
-  installNodeViaNvm,
-  isNvmInstalled,
-  resolveNodeForOxlint,
-} from "./utils/resolve-compatible-node.js";
+import { resolveConfigRootDir } from "./utils/resolve-config-root-dir.js";
 import { resolveLintIncludePaths } from "./utils/resolve-lint-include-paths.js";
 import { runKnip } from "./utils/run-knip.js";
 import { runOxlint } from "./utils/run-oxlint.js";
 import { isSpinnerSilent, setSpinnerSilent, spinner } from "./utils/spinner.js";
-
-interface ScoreBarSegments {
-  filledSegment: string;
-  emptySegment: string;
-}
-
-const SEVERITY_ORDER: Record<Diagnostic["severity"], number> = {
-  error: 0,
-  warning: 1,
-};
-
-const colorizeBySeverity = (text: string, severity: Diagnostic["severity"]): string =>
-  severity === "error" ? highlighter.error(text) : highlighter.warn(text);
-
-const sortByImportance = (diagnosticGroups: [string, Diagnostic[]][]): [string, Diagnostic[]][] =>
-  diagnosticGroups.toSorted(([, diagnosticsA], [, diagnosticsB]) => {
-    const severityDelta =
-      SEVERITY_ORDER[diagnosticsA[0].severity] - SEVERITY_ORDER[diagnosticsB[0].severity];
-    if (severityDelta !== 0) return severityDelta;
-    return diagnosticsB.length - diagnosticsA.length;
-  });
-
-const collectAffectedFiles = (diagnostics: Diagnostic[]): Set<string> =>
-  new Set(diagnostics.map((diagnostic) => diagnostic.filePath));
-
-interface VerboseSiteEntry {
-  line: number;
-  suppressionHint?: string;
-}
-
-interface CategoryDiagnosticGroup {
-  category: string;
-  diagnostics: Diagnostic[];
-  ruleGroups: [string, Diagnostic[]][];
-}
-
-const buildVerboseSiteMap = (diagnostics: Diagnostic[]): Map<string, VerboseSiteEntry[]> => {
-  const fileSites = new Map<string, VerboseSiteEntry[]>();
-  for (const diagnostic of diagnostics) {
-    const sites = fileSites.get(diagnostic.filePath) ?? [];
-    if (diagnostic.line > 0) {
-      sites.push({ line: diagnostic.line, suppressionHint: diagnostic.suppressionHint });
-    }
-    fileSites.set(diagnostic.filePath, sites);
-  }
-  return fileSites;
-};
-
-const formatSiteCountBadge = (count: number): string => (count > 1 ? `×${count}` : "");
-
-const formatIssueCount = (count: number): string => `${count} ${count === 1 ? "issue" : "issues"}`;
-
-const toRuleTitle = (ruleName: string): string => {
-  const readableRuleName = ruleName
-    .replace(/^(no|prefer|require|use)-/, "")
-    .replace(/^(nextjs|tanstack-start)-/, "")
-    .replaceAll("-", " ");
-  const title = readableRuleName.charAt(0).toUpperCase() + readableRuleName.slice(1);
-  return title.replace(/\b(css|html|url|svg|jsx|api|ua)\b/gi, (match) => match.toUpperCase());
-};
-
-const computeRuleNameColumnWidth = (ruleKeys: string[]): number => {
-  const longestRuleNameLength = ruleKeys.reduce(
-    (longest, ruleKey) => Math.max(longest, ruleKey.length),
-    0,
-  );
-  return Math.max(RULE_NAME_COLUMN_WIDTH_CHARS, longestRuleNameLength);
-};
-
-const padRuleNameToColumn = (ruleName: string, columnWidth: number): string => {
-  if (ruleName.length >= columnWidth) return ruleName;
-  return ruleName + " ".repeat(columnWidth - ruleName.length);
-};
-
-const grayLine = (text: string): void => {
-  logger.log(highlighter.gray(text));
-};
-
-const grayWrappedLine = (text: string, linePrefix: string): void => {
-  grayLine(wrapIndentedText(text, linePrefix, OUTPUT_DETAIL_WRAP_WIDTH_CHARS));
-};
-
-const printCompactRuleGroupLine = (
-  ruleKey: string,
-  ruleDiagnostics: Diagnostic[],
-  ruleNameColumnWidth: number,
-): void => {
-  const firstDiagnostic = ruleDiagnostics[0];
-  const severitySymbol = firstDiagnostic.severity === "error" ? "✗" : "⚠";
-  const icon = colorizeBySeverity(severitySymbol, firstDiagnostic.severity);
-  const siteCountBadge = formatSiteCountBadge(ruleDiagnostics.length);
-  const ruleNameRendering =
-    siteCountBadge.length > 0
-      ? colorizeBySeverity(
-          padRuleNameToColumn(ruleKey, ruleNameColumnWidth),
-          firstDiagnostic.severity,
-        )
-      : colorizeBySeverity(ruleKey, firstDiagnostic.severity);
-  const trailingBadge = siteCountBadge.length > 0 ? ` ${highlighter.gray(siteCountBadge)}` : "";
-  logger.log(`  ${icon} ${ruleNameRendering}${trailingBadge}`);
-};
-
-const getWorstSeverity = (diagnostics: Diagnostic[]): Diagnostic["severity"] =>
-  diagnostics.some((diagnostic) => diagnostic.severity === "error") ? "error" : "warning";
-
-const buildCategoryDiagnosticGroups = (diagnostics: Diagnostic[]): CategoryDiagnosticGroup[] => {
-  const categoryGroups = groupBy(diagnostics, (diagnostic) => diagnostic.category);
-  return [...categoryGroups.entries()]
-    .map(([category, categoryDiagnostics]) => {
-      const ruleGroups = groupBy(
-        categoryDiagnostics,
-        (diagnostic) => `${diagnostic.plugin}/${diagnostic.rule}`,
-      );
-      return {
-        category,
-        diagnostics: categoryDiagnostics,
-        ruleGroups: sortByImportance([...ruleGroups.entries()]),
-      };
-    })
-    .toSorted((categoryGroupA, categoryGroupB) => {
-      const severityDelta =
-        SEVERITY_ORDER[getWorstSeverity(categoryGroupA.diagnostics)] -
-        SEVERITY_ORDER[getWorstSeverity(categoryGroupB.diagnostics)];
-      if (severityDelta !== 0) return severityDelta;
-      if (categoryGroupA.diagnostics.length !== categoryGroupB.diagnostics.length) {
-        return categoryGroupB.diagnostics.length - categoryGroupA.diagnostics.length;
-      }
-      return categoryGroupA.category.localeCompare(categoryGroupB.category);
-    });
-};
-
-const printDefaultRuleGroup = (
-  ruleKey: string,
-  ruleDiagnostics: Diagnostic[],
-  rootDirectory: string,
-): void => {
-  const firstDiagnostic = ruleDiagnostics[0];
-  const ruleTitle = toRuleTitle(firstDiagnostic.rule);
-  const severitySymbol = firstDiagnostic.severity === "error" ? "✗" : "⚠";
-  const icon = colorizeBySeverity(severitySymbol, firstDiagnostic.severity);
-  const siteCountBadge = formatSiteCountBadge(ruleDiagnostics.length);
-  const trailingBadge = siteCountBadge.length > 0 ? ` ${highlighter.gray(siteCountBadge)}` : "";
-
-  logger.log(`  ${icon} ${ruleTitle}${trailingBadge}`);
-  grayWrappedLine(firstDiagnostic.message, "    ");
-  if (firstDiagnostic.help) {
-    grayWrappedLine(firstDiagnostic.help, "    ");
-  }
-  if (firstDiagnostic.url) {
-    grayLine(`    ${firstDiagnostic.url}`);
-  }
-  const firstLocation = ruleDiagnostics.find((diagnostic) => diagnostic.line > 0);
-  if (firstLocation) {
-    const locationPath = toRelativePath(firstLocation.filePath, rootDirectory);
-    grayLine(`    ${locationPath}:${firstLocation.line}`);
-  }
-};
-
-const printDefaultCategoryGroup = (
-  categoryGroup: CategoryDiagnosticGroup,
-  visibleRuleGroups: [string, Diagnostic[]][],
-  rootDirectory: string,
-): void => {
-  const issueCount = formatIssueCount(categoryGroup.diagnostics.length);
-  logger.log(`${highlighter.bold(categoryGroup.category)} ${highlighter.dim(issueCount)}`);
-  for (const [ruleKey, ruleDiagnostics] of visibleRuleGroups) {
-    printDefaultRuleGroup(ruleKey, ruleDiagnostics, rootDirectory);
-  }
-  logger.break();
-};
-
-const printVerboseRuleGroup = (
-  ruleKey: string,
-  ruleDiagnostics: Diagnostic[],
-  ruleNameColumnWidth: number,
-): void => {
-  printCompactRuleGroupLine(ruleKey, ruleDiagnostics, ruleNameColumnWidth);
-  const firstDiagnostic = ruleDiagnostics[0];
-  grayLine(indentMultilineText(firstDiagnostic.message, "      "));
-  if (firstDiagnostic.help) {
-    grayLine(indentMultilineText(`→ ${firstDiagnostic.help}`, "      "));
-  }
-  const fileSites = buildVerboseSiteMap(ruleDiagnostics);
-  for (const [filePath, sites] of fileSites) {
-    if (sites.length > 0) {
-      for (const site of sites) {
-        grayLine(`      ${filePath}:${site.line}`);
-        if (site.suppressionHint) {
-          grayLine(`        ↳ ${site.suppressionHint}`);
-        }
-      }
-    } else {
-      grayLine(`      ${filePath}`);
-    }
-  }
-  logger.break();
-};
-
-const printDefaultDiagnostics = (diagnostics: Diagnostic[], rootDirectory: string): void => {
-  const categoryGroups = buildCategoryDiagnosticGroups(diagnostics);
-  const hiddenRuleGroups: [string, Diagnostic[]][] = [];
-  const visibleCategoryGroups = categoryGroups.slice(0, MAX_CATEGORY_GROUPS_SHOWN_NON_VERBOSE);
-  const hiddenCategoryGroups = categoryGroups.slice(MAX_CATEGORY_GROUPS_SHOWN_NON_VERBOSE);
-
-  for (const categoryGroup of visibleCategoryGroups) {
-    const visibleRuleGroups = categoryGroup.ruleGroups.slice(
-      0,
-      MAX_RULE_GROUPS_PER_CATEGORY_NON_VERBOSE,
-    );
-    const remainingRuleGroups = categoryGroup.ruleGroups.slice(
-      MAX_RULE_GROUPS_PER_CATEGORY_NON_VERBOSE,
-    );
-    printDefaultCategoryGroup(categoryGroup, visibleRuleGroups, rootDirectory);
-    hiddenRuleGroups.push(...remainingRuleGroups);
-  }
-  hiddenRuleGroups.push(
-    ...hiddenCategoryGroups.flatMap((categoryGroup) => categoryGroup.ruleGroups),
-  );
-
-  if (hiddenRuleGroups.length > 0) {
-    printHiddenDiagnosticsSummary(hiddenRuleGroups);
-  }
-};
-
-const printDiagnostics = (
-  diagnostics: Diagnostic[],
-  isVerbose: boolean,
-  rootDirectory: string,
-): void => {
-  if (!isVerbose) {
-    printDefaultDiagnostics(diagnostics, rootDirectory);
-    return;
-  }
-
-  const ruleGroups = groupBy(
-    diagnostics,
-    (diagnostic) => `${diagnostic.plugin}/${diagnostic.rule}`,
-  );
-  const sortedRuleGroups = sortByImportance([...ruleGroups.entries()]);
-  const visibleRuleGroups = sortedRuleGroups;
-
-  const ruleNameColumnWidth = computeRuleNameColumnWidth(
-    visibleRuleGroups.map(([ruleKey]) => ruleKey),
-  );
-
-  visibleRuleGroups.forEach(([ruleKey, ruleDiagnostics]) => {
-    printVerboseRuleGroup(ruleKey, ruleDiagnostics, ruleNameColumnWidth);
-  });
-};
-
-const printHiddenDiagnosticsSummary = (hiddenRuleGroups: [string, Diagnostic[]][]): void => {
-  const hiddenDiagnostics = hiddenRuleGroups.flatMap(([, ruleDiagnostics]) => ruleDiagnostics);
-  const renderedParts = buildHiddenDiagnosticsSummary(hiddenDiagnostics).map((part) => {
-    const [icon, ...labelParts] = part.text.split(" ");
-    return `${colorizeBySeverity(icon, part.severity)} ${highlighter.dim(labelParts.join(" "))}`;
-  });
-
-  logger.log(`  ${renderedParts.join("  ")}`);
-  grayLine("    Run `npx react-doctor@latest . --verbose` to get all details");
-  logger.break();
-};
-
-const formatElapsedTime = (elapsedMilliseconds: number): string => {
-  if (elapsedMilliseconds < MILLISECONDS_PER_SECOND) {
-    return `${Math.round(elapsedMilliseconds)}ms`;
-  }
-  return `${(elapsedMilliseconds / MILLISECONDS_PER_SECOND).toFixed(1)}s`;
-};
-
-const formatRuleSummary = (ruleKey: string, ruleDiagnostics: Diagnostic[]): string => {
-  const firstDiagnostic = ruleDiagnostics[0];
-
-  const sections = [
-    `Rule: ${ruleKey}`,
-    `Severity: ${firstDiagnostic.severity}`,
-    `Category: ${firstDiagnostic.category}`,
-    `Count: ${ruleDiagnostics.length}`,
-    "",
-    firstDiagnostic.message,
-  ];
-
-  if (firstDiagnostic.help) {
-    sections.push("", `Suggestion: ${firstDiagnostic.help}`);
-  }
-  if (firstDiagnostic.url) {
-    sections.push("", `Docs: ${firstDiagnostic.url}`);
-  }
-
-  sections.push("", "Files:");
-  const fileSites = buildVerboseSiteMap(ruleDiagnostics);
-  for (const [filePath, sites] of fileSites) {
-    if (sites.length > 0) {
-      for (const site of sites) {
-        sections.push(`  ${filePath}:${site.line}`);
-        if (site.suppressionHint) {
-          sections.push(`    ${site.suppressionHint}`);
-        }
-      }
-    } else {
-      sections.push(`  ${filePath}`);
-    }
-  }
-
-  return sections.join("\n") + "\n";
-};
-
-const writeDiagnosticsDirectory = (diagnostics: Diagnostic[]): string => {
-  const outputDirectory = join(tmpdir(), `react-doctor-${randomUUID()}`);
-  mkdirSync(outputDirectory, { recursive: true });
-
-  const ruleGroups = groupBy(
-    diagnostics,
-    (diagnostic) => `${diagnostic.plugin}/${diagnostic.rule}`,
-  );
-  const sortedRuleGroups = sortByImportance([...ruleGroups.entries()]);
-
-  for (const [ruleKey, ruleDiagnostics] of sortedRuleGroups) {
-    const fileName = ruleKey.replace(/\//g, "--") + ".txt";
-    writeFileSync(join(outputDirectory, fileName), formatRuleSummary(ruleKey, ruleDiagnostics));
-  }
-
-  writeFileSync(join(outputDirectory, "diagnostics.json"), JSON.stringify(diagnostics));
-
-  return outputDirectory;
-};
-
-const buildScoreBarSegments = (score: number): ScoreBarSegments => {
-  const filledCount = Math.round((score / PERFECT_SCORE) * SCORE_BAR_WIDTH_CHARS);
-  const emptyCount = SCORE_BAR_WIDTH_CHARS - filledCount;
-
-  return {
-    filledSegment: "█".repeat(filledCount),
-    emptySegment: "░".repeat(emptyCount),
-  };
-};
-
-const buildScoreBar = (score: number): string => {
-  const { filledSegment, emptySegment } = buildScoreBarSegments(score);
-  return colorizeByScore(filledSegment, score) + highlighter.dim(emptySegment);
-};
-
-const getDoctorFace = (score: number): string[] => {
-  if (score >= SCORE_GOOD_THRESHOLD) return ["◠ ◠", " ▽ "];
-  if (score >= SCORE_OK_THRESHOLD) return ["• •", " ─ "];
-  return ["x x", " ▽ "];
-};
-
-const BRANDING_LINE = `React Doctor ${highlighter.dim("(www.react.doctor)")}`;
-
-const buildFaceRenderedLines = (score: number): string[] => {
-  const [eyes, mouth] = getDoctorFace(score);
-  const colorize = (text: string) => colorizeByScore(text, score);
-  return ["┌─────┐", `│ ${eyes} │`, `│ ${mouth} │`, "└─────┘"].map(colorize);
-};
-
-const printScoreHeader = (scoreResult: ScoreResult): void => {
-  const renderedFaceLines = buildFaceRenderedLines(scoreResult.score);
-
-  const scoreNumber = colorizeByScore(`${scoreResult.score}`, scoreResult.score);
-  const scoreLabel = colorizeByScore(scoreResult.label, scoreResult.score);
-  const scoreLine = `${scoreNumber} ${highlighter.dim(`/ ${PERFECT_SCORE}`)} ${scoreLabel}`;
-  const scoreBarLine = buildScoreBar(scoreResult.score);
-
-  const rightColumnLines = [scoreLine, scoreBarLine, BRANDING_LINE, ""];
-
-  for (let lineIndex = 0; lineIndex < renderedFaceLines.length; lineIndex += 1) {
-    const rightColumnContent = rightColumnLines[lineIndex] ?? "";
-    const separator = rightColumnContent.length > 0 ? "  " : "";
-    logger.log(`  ${renderedFaceLines[lineIndex]}${separator}${rightColumnContent}`);
-  }
-
-  logger.break();
-};
-
-const printBrandingOnlyHeader = (): void => {
-  logger.log(`  ${BRANDING_LINE}`);
-  logger.break();
-};
-
-const printNoScoreHeader = (noScoreMessage: string): void => {
-  logger.log(`  ${BRANDING_LINE}`);
-  logger.log(`  ${highlighter.gray(noScoreMessage)}`);
-  logger.break();
-};
-
-const buildShareUrl = (
-  diagnostics: Diagnostic[],
-  scoreResult: ScoreResult | null,
-  projectName: string,
-): string => {
-  const errorCount = diagnostics.filter((diagnostic) => diagnostic.severity === "error").length;
-  const warningCount = diagnostics.filter((diagnostic) => diagnostic.severity === "warning").length;
-  const affectedFileCount = collectAffectedFiles(diagnostics).size;
-
-  const params = new URLSearchParams();
-  params.set("p", projectName);
-  if (scoreResult) params.set("s", String(scoreResult.score));
-  if (errorCount > 0) params.set("e", String(errorCount));
-  if (warningCount > 0) params.set("w", String(warningCount));
-  if (affectedFileCount > 0) params.set("f", String(affectedFileCount));
-
-  return `${SHARE_BASE_URL}?${params.toString()}`;
-};
-
-const printCountsSummaryLine = (
-  diagnostics: Diagnostic[],
-  totalSourceFileCount: number,
-  elapsedMilliseconds: number,
-): void => {
-  const errorCount = diagnostics.filter((diagnostic) => diagnostic.severity === "error").length;
-  const warningCount = diagnostics.filter((diagnostic) => diagnostic.severity === "warning").length;
-  const affectedFileCount = collectAffectedFiles(diagnostics).size;
-  const totalIssueCount = diagnostics.length;
-  const elapsedTimeLabel = formatElapsedTime(elapsedMilliseconds);
-
-  const issueCountColor =
-    errorCount > 0 ? highlighter.error : warningCount > 0 ? highlighter.warn : highlighter.dim;
-  const issueCountText = `${totalIssueCount} ${totalIssueCount === 1 ? "issue" : "issues"}`;
-  const fileCountText =
-    totalSourceFileCount > 0
-      ? `across ${affectedFileCount}/${totalSourceFileCount} files`
-      : `across ${affectedFileCount} file${affectedFileCount === 1 ? "" : "s"}`;
-  const elapsedTimeText = `in ${elapsedTimeLabel}`;
-
-  logger.log(
-    `  ${issueCountColor(issueCountText)} ${highlighter.dim(`${fileCountText}  ${elapsedTimeText}`)}`,
-  );
-};
-
-const printSummary = (
-  diagnostics: Diagnostic[],
-  elapsedMilliseconds: number,
-  scoreResult: ScoreResult | null,
-  projectName: string,
-  totalSourceFileCount: number,
-  noScoreMessage: string,
-  isOffline: boolean,
-): void => {
-  if (scoreResult) {
-    printScoreHeader(scoreResult);
-  } else {
-    printNoScoreHeader(noScoreMessage);
-  }
-
-  printCountsSummaryLine(diagnostics, totalSourceFileCount, elapsedMilliseconds);
-
-  try {
-    const diagnosticsDirectory = writeDiagnosticsDirectory(diagnostics);
-    logger.log(highlighter.gray(`  Full diagnostics written to ${diagnosticsDirectory}`));
-  } catch {
-    /* swallow — failing to write the dump shouldn't block the summary */
-  }
-
-  if (!isOffline) {
-    logger.break();
-    const shareUrl = buildShareUrl(diagnostics, scoreResult, projectName);
-    logger.log(`  ${highlighter.bold("→ Share your results:")} ${highlighter.info(shareUrl)}`);
-  }
-};
-
-const resolveOxlintNode = async (
-  isLintEnabled: boolean,
-  isQuiet: boolean,
-): Promise<string | null> => {
-  if (!isLintEnabled) return null;
-
-  const nodeResolution = resolveNodeForOxlint();
-
-  if (nodeResolution) {
-    if (!nodeResolution.isCurrentNode && !isQuiet) {
-      logger.warn(
-        `Node ${process.version} is unsupported by oxlint. Using Node ${nodeResolution.version} from nvm.`,
-      );
-      logger.break();
-    }
-    return nodeResolution.binaryPath;
-  }
-
-  if (isQuiet) return null;
-
-  logger.warn(
-    `Node ${process.version} is not compatible with oxlint (requires ${OXLINT_NODE_REQUIREMENT}). Lint checks will be skipped.`,
-  );
-
-  if (isNvmInstalled() && process.stdin.isTTY) {
-    const { shouldInstallNode } = await prompts({
-      type: "confirm",
-      name: "shouldInstallNode",
-      message: `Install Node ${OXLINT_RECOMMENDED_NODE_MAJOR} via nvm to enable lint checks?`,
-      initial: true,
-    });
-
-    if (shouldInstallNode) {
-      logger.break();
-      const freshResolution = installNodeViaNvm() ? resolveNodeForOxlint() : null;
-      if (freshResolution) {
-        logger.break();
-        logger.success(`Node ${freshResolution.version} installed. Using it for lint checks.`);
-        logger.break();
-        return freshResolution.binaryPath;
-      }
-      logger.break();
-      logger.warn("Failed to install Node via nvm. Skipping lint checks.");
-      logger.break();
-      return null;
-    }
-  } else if (isNvmInstalled()) {
-    logger.dim(`  Run: nvm install ${OXLINT_RECOMMENDED_NODE_MAJOR}`);
-  } else {
-    logger.dim(
-      `  Install nvm (https://github.com/nvm-sh/nvm) and run: nvm install ${OXLINT_RECOMMENDED_NODE_MAJOR}`,
-    );
-  }
-
-  logger.break();
-  return null;
-};
 
 interface ResolvedScanOptions {
   lint: boolean;
@@ -617,51 +76,6 @@ const mergeScanOptions = (
   adoptExistingLintConfig: userConfig?.adoptExistingLintConfig ?? true,
   ignoredTags: buildIgnoredTags(userConfig),
 });
-
-const printProjectDetection = (
-  projectInfo: ProjectInfo,
-  userConfig: ReactDoctorConfig | null,
-  isDiffMode: boolean,
-  includePaths: string[],
-  lintSourceFileCount?: number,
-): void => {
-  const frameworkLabel = formatFrameworkName(projectInfo.framework);
-  const languageLabel = projectInfo.hasTypeScript ? "TypeScript" : "JavaScript";
-
-  const completeStep = (message: string) => {
-    spinner(message).start().succeed(message);
-  };
-
-  completeStep(`Detecting framework. Found ${highlighter.info(frameworkLabel)}.`);
-  completeStep(
-    `Detecting React version. Found ${highlighter.info(`React ${projectInfo.reactVersion}`)}.`,
-  );
-  completeStep(
-    `Detecting Tailwind. ${
-      projectInfo.tailwindVersion
-        ? `Found ${highlighter.info(`Tailwind ${projectInfo.tailwindVersion}`)}.`
-        : "Not found."
-    }`,
-  );
-  completeStep(`Detecting language. Found ${highlighter.info(languageLabel)}.`);
-  completeStep(
-    `Detecting React Compiler. ${projectInfo.hasReactCompiler ? highlighter.info("Found React Compiler.") : "Not found."}`,
-  );
-
-  if (isDiffMode) {
-    completeStep(`Scanning ${highlighter.info(`${includePaths.length}`)} changed source files.`);
-  } else {
-    completeStep(
-      `Found ${highlighter.info(`${lintSourceFileCount ?? projectInfo.sourceFileCount}`)} source files.`,
-    );
-  }
-
-  if (userConfig) {
-    completeStep(`Loaded ${highlighter.info("react-doctor config")}.`);
-  }
-
-  logger.break();
-};
 
 export const scan = async (
   directory: string,
